+++ django_mysql/models/query.py
@@ -141,7 +141,7 @@ class ApproximateInt(int):
 
 class SmartChunkedIterator(object):
     def __init__(self, queryset, atomically=True, status_thresholds=None,
-                 pk_range=None, chunk_time=0.5, chunk_max=10000,
+                 pk_range=None, chunk_time=0.5, chunk_size=2, chunk_max=10000,
                  report_progress=False, total=None):
         self.queryset = self.sanitize_queryset(queryset)
 
@@ -156,8 +156,8 @@ class SmartChunkedIterator(object):
         self.pk_range = pk_range
 
         self.rate = WeightedAverageRate(chunk_time)
-        self.chunk_size = 2  # Small but will expand rapidly anyhow
         self.chunk_max = chunk_max
+        self.chunk_size = min(chunk_size, chunk_max)
 
         self.report_progress = report_progress
         self.total = total