+++ django_mysql/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = 'Adam Johnson'
 __email__ = 'me@adamj.eu'
-__version__ = '0.2.1'
+__version__ = '0.2.2'
 
 
 default_app_config = 'django_mysql.apps.MySQLConfig'